use only unsigned numbers in md5::buildverts()

+++ src/engine/model/md5.cpp
@@ -478,7 +478,7 @@ void md5::md5mesh::buildverts(const std::vector<md5joint> &joints)
     {
         md5vert &v = vertinfo[i];
         vec pos(0, 0, 0);
-        for(int k = 0; k < v.count; ++k)
+        for(ushort k = 0; k < v.count; ++k)
         {
             const md5weight &w = weightinfo[v.start+k];
             const md5joint &j = joints[w.joint];
@@ -493,7 +493,7 @@ void md5::md5mesh::buildverts(const std::vector<md5joint> &joints)
 
         blendcombo c;
         int sorted = 0;
-        for(int j = 0; j < v.count; ++j)
+        for(ushort j = 0; j < v.count; ++j)
         {
             const md5weight &w = weightinfo[v.start+j];
             sorted = c.addweight(sorted, w.bias, w.joint);
